[FEATURE] Add possibility to request a run level
It might happen that a command must do some work
in a lower run level but later on needs a higher
level. This is e.g. the case for the cache command
which needs low level access to be more robust,
but later needs to know about caches in extension
configuration which is only available in higher
run levels.