[!!!][TASK] Remove compatibility code for TYPO3 6.2
As the version of typo3_console in this branch is not
compatible with TYPO3 6.2 anyway, we can remove all
compatibility code related to that TYPO3 version.

Although this should not break something in third party
code, this is a rather extensive removal, thus marked
as breaking. Check your commands or forks!