[!!!][FEATURE] Load active packages configuration from environment
!WARNING! This change breaks your deployment scripts, if you rely on
the install:generatepackagestates command!

While previously the configuration was taken from the composer.json file,
it is now taken from the environment variable TYPO3_ACTIVE_FRAMEWORK_EXTENSIONS
which must hold a comma separated extension key list.

This is far more flexible and much more reliable, as we can activate different
extensions in different scenarios and do not need to parse the composer.json.

Additionally, *ALL* extensions which end up in 'typo3conf/ext' will be *ACTIVE*
by default! Thus, you only need to add system extensions which are
delivered in the typo3/cms package to the configuration.

This will be released as part of 2.0 soon.