[FEATRUE] Let the newly created plugin copy the console
Instead of putting the burden upon users to add a script section
to the composer.json, we have a plugin now, which takes care
of copying the script.

At a later point we may even decide to do more things during
composer installation, we might otoh decide to remove the plugin again.

For now it is a useful feature worth being created.