[FEATURE] Add database import/export commands
This adds the feature like requested in #93 and suggested in #95
but with the conceptual change, that imports and exports are done
over stdin and stdout instead of using files.

This adds much more flexibility and is secure by default, as the output
is never stored in a file, which could be compromised.

This is also a small step to get #125 resolved.