[BUGFIX] Fail if existing DB is selected but not empty
If during setup the user decides to use and existing DB,
but the DB is not empty, the installation must not fail with a fatal error.

The installation process is now streamlined, that a check whether the
execution of one installation step needs to be executed is performed
before the user input is requested and after the actin was performed,
so that we can now output an error message if the action failed.