[TASK] Streamline and document backend restrictions
The backend restriction commands are now streamlined to be more understandable.
Besides that, a documentation is added on top of the class.

Goal is to document every command controller class like that prior to 3.0 release.

This is a breaking change only for early adopters of the 3.0.x-dev branch,
as the editor lock commands were not available brforehand.