[FEATURE] Make it possible to disable binary copy to root
As composer now can take care to copy (or link) the binaries,
we can add an option to not additionally copy the binary
to the root directory.

This option is "install-binary" in the plugin config
and is on by default so that third party libraries can adopt the new location.