[TASK] Make console a composer library
This is not breaking! The console extension will exist
and will be distributed through TER to make non composer users happy.
Required extension files will be generated right before uploading
to TER.

Cloning this repository will for a while deliver TYPO3 extension
files for the time being, but it is recommended to switch to other
distribution ways, preferably composer installation.