[BUGFIX] Do not show cache:flush when T3 not installed
cache:flush command needs to be in low runlevel
to not load extension files before clearing caches.
However it is also not useful when TYPO3 is not installed yet.

Thus we just hide it until we have a nicer bootstrap,
TYPO3 does not need extension caching any more or some
other nicer solution magically appears.

Fixes #281