[BUGFIX] Abort installation if LocalConfiguration.php exists
TYPO3 cannot reliably install if this file exists. There are
way to many checks for that file, so that if it exists,
executing installation code will lead to unpredictable results.

Therefore we now abort the installation in non interactive mode
and offer the deletion of this file in interactive mode.

Fixes #282