[BUGFIX] Do not prefix fields/tables on "*.change" DB schema update
The prefixing of fields and tables is a basically destructive schema
update, thus it is not really expected when "*.change" updates are
requested.

Fix this by splitting update suggestions into safe and destructive
statements and introducing the "field.prefix" and "table.prefix"
schema update types.

This is mildly breaking if you expect field/table prefixing upon
"*.change" but it actually ensures that your data is left untouched
unless requested otherwise.

Fixes #301