[BUGFIX] Install extension dummy by default
Currently extension can bring their own commands, but with
TYPO3 Console not being an extension any more, these cannot
depend on it. Also other tools like Surf look for the console
in the extension directory.

Until things are sorted out, we add the possibility to add
an extension dummy in the extension directory to make dependent
extensions and external tools happy.

This option will be on by default but deprecated in master and not
deprecated in 3.x

Fixes #298