[BUGFIX] Call data handler API when flushing caches
Because of the removal of the realurl cache table clearing workaround,
we must now take care to call the data handler API, so that realurl
hooks subscribers that clear the caches are called.

All other cache hook subscribers are thus called as well.

Nevertheless we consider this legacy API and mark the according methods
as deprecated and internal at the same time.
Nothing will be exposed to userland.

Fixes #286