[!!!][FEATURE] Auto-create _cli_lowlevel user (#339)
Instead of creating the (unfortunately still required) cli user
during installation, we now create it during runtime,
if it does not exist.

This has the benefit, that we have this code in one place
and have it simplified.

This comes with a little downside, that if a disabled user already
exists, that it cannot be created any more. We accept this as the
likeliness that this happens is rather low.

This change is only slightly breaking because we at the same
time removed the MCONF name global.