[!!!][FEATURE] Add "all.safe" to database:updateschema
Refactor the schema types to be also responsible for the
mapping to the statement types.

This has the little downside that this class now has a
dependency to internal array structure of TYPO3 schema updates,
however it is natural from a users perspective to have the
new types in this enum as well.

We now also added tests for this class and fixed some
issues found due to these tests.

Also add new documentation to explain the new possibilities.

On top of that, the command has the non destructive operations
as default so that it can be called without arguments.