[BUGFIX] Avoid wrong and unnecessary database schema changes
When calling the extension:setupactive command, the console
called TYPO3 API that is usually used when activating extensions.
However this API is called also for extensions, that are already active.

This leads to the situation, that the database schema is changed to what
the individual extension requires. If however another extension is active, that
is based on the currently activated extension, the correct schema is reverted
to state without the depending extension. Although this is fixed again, when the depending extension is processed, these operations impose a lot of overhead.

Even worse, these operations can be destructive, when the schema
changes are destructive. This is e.g. the case when a VARCHAR field is
made smaller, but already contains data exceeding the new smaller field size.

We now do all non database operations on all extensions first, do a
database schema adjustment once taking all extensions into account and lastly
import database tables or t3d files of every extension if needed.

This not only fixes the bug mentioned above, but speeds up the
execution time for this command by a lot, especially in installations
with many active extensions.

To be consistent and to avoid user errors with potential data loss,
we change the extension:activate command as well accordingly.

Fixes #382