[BUGFIX] Reload caches upon extension activation
When switching to our own API for activating extensions,
it was missed that during activation is is necessary to flush
caches before performing setup operations.

Besides that, the signal sent when activating an extension
also was omitted.

So we now manually reload and flush caches and emit the signal
when required.

Fixes #385