[BUGFIX] Remove table.clear update type
This operation is implemented in TYPO3 versions lower 8.4
and was meant to trigger a TRUNCATE TABLE in the following
situations:

1. A cache table needs to be cleared
2. An auto_increment field is added, but not the key

However the code was never executed, as these operations are not part
of the install tool database compare action.

Unfortunately typo3_console revived this previously dead code and reveals
a fatal error.

When the auto_increment field exists, but needs a change
(e.g. from INT(11) to INT(10)), then a truncate of these tables is forced.

Since these can be ANY table, it might destroy your data.

Therefore we remove this operation again from typo3_console and let
the code behind it rot in peace.

Fixes: #391