[FEATURE] Always generate PackageStates.php file (#389)
Although it is pretty easy and convenient to add install:generatepackagestates
as command to be run after composer dumped its autoload information,
this isn't yet adopted widely. To change that, we use our plugin
to do so, so that the PackageStates.php file is now always written
on composer invocation.

Hopefully people using composer will eventually stop fiddling
with the extension:activate commands as of the next version of TYPO3 Console

Besides that we remove the broken code from this command, which triggered
removal of directories of inactive extensions and add a dedicated (working)
command for that.