[FEATURE] Enable XSD generating via typo3/fluid-schema-generator
This change adds typo3/fluid-schema-generator as
dependency and rewrites the documentation:generatexsd
command so it uses the SchemaGenerator from that
package. The benefit is that the schema generator now
becomes capable of creating schemas based on multiple
namespaces, which is required in order to generate a
schema on TYPO3v8+.

The specifics are:

* ClassLoader is made available via the ConsoleBootstrap
   using a public method. This is required in order to fetch
   the PHP namespace to folder path mapping.
* Input arguments for documentation:generatexsd command
   reduced to a single array of PHP namespaces, where
   multiple namespaces can be specified as CSV.

The XsdGenerator class is now redundant - and removed.