[BUGFIX] Correctly pass boolean arguments to the sub process
The CLI RequestBuilder maps boolean arguments to a bool, but
does not consider an empty string '' to be a boolean false.

Because such arguments are already converted to boolean in the main
process, we must therefore convert this boolean value to a
string that is then correctly understood by the RequestBuilder as boolean
value.