[BUGFIX] Make compatible with TYPO3 8LTS
Due to some changes in the install process it has become
apparent, that we need to execute the silent configuration
upgrade more often, likely directly after an action was executed.

For that we move this into the child process.

This makes the install work again, but with an error message, that
hints to to the second required fix, which is providing the database
driver.

For now we hard code it to mysqli and make it available as option
in later releases.

Fixes: #448
Closes: #449