[BUGFIX] Do not write extConf if present already
When using extension:setupactive there are several steps
performed for each extension. One step is writing the
default extension configuration to LocalConfiguration.php
This is done regardless there is a configuration present already.

To avoid having the need to have write access to LocalConfiguration.php
in production, this step is now skipped when extConf for the
extension is already present for extensions that are set up
using extension:setupactive