[FEATURE] Upgrade Automation
Add new upgrade:* commands to list necessary upgrade steps,
run through all the steps, or execute single steps out of the list.

To achieve this, a bunch of classes are added with a clean,
stable and straightforward API.

The commands handle the upgrade steps, but (of course)
do NOT perform the code update itself. This still needs
to be done manually (e.g. via Composer) beforehand.

So basically these commands are a replacement for the (manual)
click-through in the TYPO3 Install Tool AFTER code update.