[TASK] Move functional tests to PHPUnit
Use command dispatcher now to execute the commands
and plain PHPUnit to compare the result with the expected result.

This is especially great for Appveyor, as we now have all tests
executed on this platform as well.

Besides that the structure for functional tests is much clearer
and more streamlined.

This change revealed some minor inconsistencies in some commands,
which are fixed as well.