[FEATURE] Deactivate not compatible extensions on upgrade
Before preforming any upgrade command, we now check whether
extensions are found that are marked as incompatible, as well
as individually checking if the basic extension bootstrap
would work.

Any incompatible extension found will be deactivated during this process.

This ensures that the upgrade can complete even if some incompatible
extensions are present in the system.