[BUGFIX] Only activate default exts in non composer mode
Also ensure that a package states file exists that reflects
the defined set of system extensions, but no third party extensions,
to avoid interference during the setup.

Besides that in the end re-create the package states file again,
flush caches and set up all extensions properly.

The end result will be a clean and desired state.