[BUGFIX] Allow to update DB schema without ext_tables.php
We added this step to the bootstrap of
database:updateschmea command only because TYPO3 8 needs
TCA to add additional fields to the schema.

With this we however run into a chicken/egg issue, when an
extension tries to access DB (most likely a cache table)
in the process of reading ext_tables.php

Now we read TCA early on, when reading extension configuration,
which is exactly what happens in TYPO3 9 anyway,
so we can skip reading ext_tables.php for updating the DB schema.