[BUGFIX] Always exclude extensions when requested
For generating PackageStates.php we now mark
local extensions as inactive when they are excluded,
even if they have been marked as protected by the
extension owner.

We now also allow framework extensions to be excluded,
but only if they are neither protected, nor provided
in the list of active framework extensions.

This makes it possible to exclude some framework extensions,
when using the --activate-default switch.

Fixes #541