[FEATURE] Add --skip-extension-setup to install:setup (#572)
Allow to skip extension setup. This allows to make sure
installation succeeds.

While using this switch avoids issues with extensions,
it will leave the installation in a weird state,
with all third party extensions not active and
some core extensions that might not be needed in the
project being active.