[!!!] Remove deprecated code
Finally the dummy extension is not installed any more,
so the extra section in root composer.json can be removed.

Generating package states code that was hidden behind
a feature flag and deprecated has been removed.

During install:setup dependencies to not required core
extensions will lead to exceptions. The extensions or
the project setup needs to be fixed.