[!!!] Remove own autocomplete implementation
Remove our autocomplete implementation in favor of:
https://github.com/bamarni/symfony-console-autocomplete

Since we use a regular Symfony console application,
we can rely on existing tools instead of
having to maintain our own code.

This means however, that autocomplete support
is removed for TYPO3 7.6 because with this
TYPO3 version, we cannot use a Symfony application
because as it comes with a too old symfony/console
version.