[!!!] Back to the roots. Remove commands
… that are now implemented in a sane way in TYPO3.
Our goal always was to rather provide a good
infrastructure than implementing a bunch of commands.

Although some of the removed commands have some benefits in
some aspects over the ones from the core, but that is
still acceptable. We could port some of the benefits
to TYPO3 so that eventually everything is going to
be awesome.