[FEATURE] Allow replacing commands from other packages
It is now possible to override/ replace/ remove
commands from other packages (or even the core
or TYPO3 Console themselves).

A package can define a "replace" section in the
configuration with an array of command identifiers.
Commands having these identifiers will be skipped
in any case, whether they really have a replacement
or not.