Fix upgrade handling of extension install wizards (#613)
Instead of hard coding a list of wizards that install
extensions, we now check if the wizard extends the
abstract class. By that we now track more wizards,
which were previously missing in the hard coded list.

Also fix getting the wizard object in this process
for non composer mode.

Last but not least, add warning messages in composer mode
to notify users that extensions have not been installed,
thus might be lacking after the upgrade.