[!!!][TASK] Undo rename of binary (#637)
While it is desirable to have the same binary name with or without
TYPO3 Console installed when doing cli actions in TYPO3 projects,
this time has not come yet. TYPO3 9.0.0 is released,
still using "typo3" as command name and most likely it will
stay like that until TYPO3 v10.

Even if "typo3console" would be a much superior name,
(which in fact isn't) it wouldn't justify to break
such an essential thing for upgraders at this point in time.

I'm sorry for those of you who used dev-master already
(thanks for that btw.), but I think it is wiser to mitigate
the breakage for upgraders when switching from 4.9.x to 5.x

At the same time, we move the binary into the root folder,
which does not affect users, but will ease development a bit.