[TASK] Swap compatibility classes
Everything in Classes now is compatible with the latest
stable TYPO3 release. In case compatibility for older versions
is required a folder with differing classes is created in Compatibility.

Same for TYPO3 master compatibility. It is achieved first by
putting a folder in Compatibility and once this TYPO3 version is
released, these classes move to Classes folder.