[BUGFIX] Correctly register class loading in non composer mode
In order for the compatibility layer to work, the Compatiblity
folder must also be registered in the class loader.

We now change the non composer class loading to register
all psr-4 namespaces directly from the composer.json file,
which will avoid issues in the future.

With this change we also simplify the autoload section
in our composer.json

Fixes: #655