[BUGFIX] Do not pass literal "false" as argument value
In TYPO3 command line parsing a literal string "false" as argument
value was converted to boolean false. Since we now use
Symfony argument parsing, this does not happen any more.

The previously introduced hack now leads to the opposite
that the value is converted to boolean true.

We now remove this hack, as Symfony correctly converts an empty string
as argument value to false if the option type is bool.