[BUGFIX] Output command method return value, but deprecate it
In previous versions of typo3 console and in the pure
TYPO3 implementation the command method return value
is set as response content. This content is then later
just echo'ed out, not respecting the output object that is used
otherwise, which means formatting is never applied.

For the sake of backwards compatibility we now output the
command method result again, but properly use the output object
and at the same time deprecate this functionality, as it
isn't very helpful to implicitly output something while
it is equally simple to use the output API explicitly.

Fixes: #658