[BUGFIX] Do not call last install step twice
Because TYPO3 core installer class does not implement
the method to check whether this last step was executed
and we checked if the action needs to be executed again
right after execution and by default we assumed yes,
the action was triggered twice.

Since this action deletes the first install file,
we change the default whether the action needs to
be executed to existence of this file.

Fixes: #668