[BUGFIX] Fix command execution in composer run context
When executing console binary on Windows in composer run
context, we must not use the binary installed in the bin
folder because this is not a php script but a proxy file.

Therefore we now just use our binary directly.

This makes giving the script event to the factory method
obsolete, but we keep the API for now to avoid a breaking change.