[BUGFIX] Fix package names of TYPO3 extensions in composer mode
To have consistent package names (and thus consistent full
command identifiers), we need to calculate the extension key
from composer packages that are TYPO3 extensions and use
the extension key as package name.

Replaces can now be simplified to only use extension keys.

Since we removed the "extbase" replace, which was a command
registered in earlier TYPO3 8.7 versions, we now require
the latest 8.7 release.