[BUGFIX] Fix Symfony command registration for extensions
Remove the registration of Symfony commands for ALL TYPO3
extensions in composer mode, because it could happen
that the commands were available through TYPO3 Console
despite the core extension(s) being inactive.

In order to be able to check whether a command should be
shown, we need to ask the PackageManager whether the extension
which provides this command is active.

For that we remove the early return for all commands
gathered from composer packages during composer build time
and always evaluate if we need to add something from active extensions.

This also fixes the issue that commands from extensions that
are bound to the root package are not found in composer mode, as we delegate
finding the extension to TYPO3 instead of evaluating what
actually is installed via composer.

To clarify the new behavior, AllCommands.php is now renamed to
ComposerPackagesCommands.php

Fixes: #672