[FEATURE] Use CommandLoader API of symfony/console
Instead of instantiating all command objects upfront
before running the application, we now use a command
provider, that is creating command objects on demand.

This should improve startup time when running individual commands,
especially if commands have heavy initializations.

This also fixes a bug when TYPO3 Symfony commands were
accessing TYPO3 state in the constructor, which was not possible,
due to early instantiation.

Fixes: #675