[BUGFIX] Use Symfony application description to render reference
Instead of maintaining a separate repo for the command,
that renders the reference, we include it in our main repo now.

With this change, we rework the rendering to use a Symfony application
description instead of the command manager a base to generate
the reference.

Result is a more consistent reference, that matches the existing commands
better than before.

This also enables us to slowly migrate from command controllers
to native Symfony commands, but still have these in the reference.