[FEATURE] Rename install:generatepackagestates command
This command was put into install: namespace mainly
because it was not possible to have it in extension:
namespace due to technical constraints.

Since we now can have individual classes for each command,
we can rename it to a nicer name, which is "extension:dumpactive"

The old command name and option names still work
but will show a deprecation message.