[FEATURE] Make it possible to abort installation process
Install actions currently can throw exceptions to abort,
but we now introduce an exception that makes the installation
abort, but gracefully let's the dispatcher return false
instead of letting the exception bubble up.

We use this now to add the integrity check to the preparation step
and removing this check from the command controller.