[BUGFIX] Suppress error output for sub-processes
In order to not interfere with the actual output
of sub-processes, we suppress the native php error output
for starting a sub-process.

Any important error will be turned into an exception anyway,
which will result in the sub-process to fail.

Also fix the non working PHP_INI_PATH env var setting
since we are at it.

Fixes: #751