[BUGFIX] Fix handling of chatty upgrade wizards
The TYPO3 code introduced a ChattyInterface for upgrade
wizards in d336193d86e3c2b99da6d59870f56116277e89c3.

When one of those wizards tries to output something it crashes
because the output is not initialized.

This is fixed by using a BufferedOutput and attaching the collected
messages to the upgrade wizard result.