[BUGFIX] Always disable all caches for low level commands
Instead of picking some caches to disable and re-anable them later,
we now initialize the CacheManager with disables caches for low level commands.
Full commands now always get full caches without any further ado.
This is the same behaviour as in TYPO3 9, which we mimik also for TYPO3 8.

Further adaptions to the TYPO3 9 bootstrap will follow in later commits.

Fixes: #760