[FEATURE] Always show command list and help
When TYPO3 is set up and a LocalConfiguration.php exists,
then TYPO3 Console tried to fully bootstrap TYPO3 also
for showing the list of commands.

This lead to exceptions, e.g. when the configured database connection
wasn't working.

We now only do a preliminary boot, so that at least the command
list is shown and only fail in such cases when executing concrete commands