[FEATURE] Delay output of booting errors
Instead of failing before starting the console application,
we now catch booting errors and output them once the IO
is properly configured.

This means the following:

* Booting errors for all registered commands will lead to an exception just like before.
* list and help commands gracefully work but return exit code 255 and show a tiny error message
* list and help commands also throw the exception when output is verbose

This ensures that no command is executed in limbo state, except help and list,
which are not changing any state.

As "side effect" auto completion will still work, even if the current installation
is broken.