[FEATURE] Show more helpful messages in error conditions
First and foremost, we now always register all commands under
all circumstances. This means, that auto completion will always
work properly. The command listing will still be reduced to the commands
that are likely to work, but optionally allows showing all commands.

Also show a more helpful error message when executing a command
which cannot be run unless TYPO3 is set up instead of showing
an error message that the command does not exist (which is in fact wrong)