[FEATURE] Package complete console in TYPO3 extension
In order to ease new documentation rendering, a new way
the extension is built is introduced. The result will
be automatically pushed to a new repo, which will then
trigger the rendering on each change.

As a "side effect" some code can be simplified, as console
now also as extension is installed via composer.