[BUGFIX] Improve error and exception handling
Previously we ignored all TYPO3 configuration regarding
error and exception handling, to get a more reliable behavior
and a decent error output.

With this change we loosen this strictness a bit and at least look
at the configured error levels that should be logged or turned
into exceptions, but still enforce a reasonable configuration, by not
allowing warnings to be omitted.

Additionally we make sure both exceptions and errors are properly logged
using the logging framework.

We still disallow changing the error and exception handler to disallow
misconfiguration and also still not log errors to the legacy database log.

In case the latter is needed, it would still be possible to create a writer,
that writes to the database in a way that the TYPO3 backend log module
can use it.