Remove dependency to Composer plugin helper
Since TYPO3 9.5 there is no possibility to use the
typo3/cms mono repo for Composer installations.

Therefore we must not compile the list of to be activated
framework extensions based on the required packages any more,
because only required framework extensions will be installed
anyway. Therefore we can just activate all framework extensions
that are found in Composer mode.

Since compiling this list was the remaining purpose of the plugin,
we can remove the dependency now.