[TASK] Add option to fail on error to extensionsetupifpossible
This command is a wrapper to execute:

database:updateschema
cache:flush
extension:setupactive

Initially this was meant to be used in Composer scripts to
not interfere Composer install on build systems, but at the same
time perform necessary steps in dev systems.

It was, until now, not suitable for deployment scenarios though,
because during a deployment it is important that the command fails
when something goes wrong to avoid a broken state after a deployment.

The added option makes it now suitable for deployments, too.

Fixes: #925