[!!!][BUGFIX] Remove evaluation of Commands.php in Extensions
This change should be considered as part of 7.0.0, but was lacking.
Without it a fatal error occurs, because commands from TYPO3 Console
are registered twice.

Since TYPOO3 now can handle low level commands with TYPO3 11,
the possibility to register them with TYPO3 Console has become
obsolete and therefore is removed.

Fixes: #1009