[DOCS] Remove targets from menu
The page is currently empty. Having a "Link targets" in the menu
used to be standard. However, this was often critized. It is
non-intuitive for readers and often an explanation was not included.

Since now the page was empty anyway, it did not have a useful
purpose and is safe to remove.