[BUGFIX] Disable extension constraint checks in Composer mode
When TYPO3 and extensions are installed with Composer, it is pointless
to check for ext_emconf.php files and version constraints in them,
because they are not evaluated any more since TYPO3 11.5.

Therefore, even if those files contain outdated information,
those have no effect on the installation and as such should
not lead to any errors.

Also fix upgrade:prepare to not deactivate an extension, as this
leads to errors, as packages can not be deactivated in Composer
mode during runtime any more

Fixes #1040