[FEATURE] Add global site option
There are quite some use cases for cli commands,
that do tasks, that are bound to a dedicated site.

In order to avoid code duplication and diverging
implementations of the same thing, a global option
is added, that can be used for any command.

Implementers of commands can then retrieve the
site object directly from $input->getOption('site')
an execute any task associated with this site.

This option is completely ignored, when commands
do not evaluate it.