[TASK] Provide return types for Symfony Commands
This change adds return types,
which are required for the compatibility
with Symfony 6.

This is a pre-patch for supporting
Symfony 4-6, however a few further changes
are necessary as well to provide
Symfony 6 compat.

It should be noted that
TYPO3 v11 LTS requires Symfony 5.4, and
TYPO3 v12 requires Symfony 6.x.

See https://github.com/TYPO3/typo3/blob/11.5/composer.json