[BUGFIX] use siteBaseUrl for install steps argument definition
\Helhum\Typo3Console\Install\Action\InteractiveActionArguments::populate loops
over the argument definitions of the install step by name. The input options
from the `install:setup` command define `--site-base-url`, which is translated
to the givenArguments key `siteBaseUrl`.

Passing a base URL other than `/` to `--site-base-url` did not work at all,
because the code looked for a `siteUrl` option instead.